Adding code to pass back the username when a failed password is entered
This stops users having to retype a username if they get the password
wrong. It may not be ideal behaviour, as users may retry with incorrect
usernames. However, that shouldn't lock the account (the username is
incorrect) and is probably less annoying than having to retype usernames
all the time